refactor: move types to types/index.ts
